feat: New broadcast impl
1)完成了新broadcast的实现
2)管理broadcast的任务被分成了三个角色：GuideManager, DownloadManager, BroadcastManager
3)这三个角色启动的位置有讲究，分别是在提交端，executor(DownloadManager在这里启动),task(主要
是BroadcastManager)，对于MultiProcessScheduler的处理，差不多，但是要将pool
的初始化放在submitTasks开始的时候.
4)GuideManager用于保存所有的每个broadcast变量所对应的在集群上的存留信息，这个Manager应该在
提交进程中使用
5)DownloadManager用于从GuideManager获取source信息，并根据source信息向其他的
DownloadManager取需要的数据相应每个task的需求，从远程抓来数据，保存当前的数据
6)BroadcastManager用于对task内的数据进行登记，并将其放到DownloadManager中进行管理，所有
task中需要使用数据的时候，调用此接口
7)DownloadManager的broadcast数据块存放在workdir的目录下，每个文件在DownloadManager中用
mmap读写，bitmap保存每个块在文件中的offset与size，文件名和block的信息保存在共享内存中
供BroadcastManager获取相关的数据，下载和读取的同步由锁完成，每个锁处理一个broadcast，
下载完成后，锁直接删除相关信息见历史记录中的<6><7>
8)因为在executor中增加了broadcast var变量的元数据信息，需要将64M的内存扩大到128M

历史修改记录

<1>fix: add start_manager and move the definition according to multi process

<2>fix: To adjust the environ

<3>fix: fix the bug and let download fast
1)这个配置主要是为了处理使用共享内存操作时的一些优化措施，比如executor
保存一个本地副本，这样下载的时候时间不会太长，后来使用本地file之后，
这个问题就不复存在了
2)将DownloadManager,GuideManager的关系保持各自独立，两类manager在任务结束时
都需要自己join自身启动的thread，防止出现程序退出不干净的情况,clear相应的uuid也是
每个manager负责每个相应的数据
3)每个新的uuid需要download的时候，单独启动一个thread，这样不会影响zmq的socket收发
其他的数据
4)将fetch的大小，由原来单独一块，变成一次批发指定数目的block，本以为可以提高性能
后来发现影响不大

<4>fix: some fix and dealing with OOM

这里用Condition对BroadcastManager和DownloadManager之间的数据做了一个同步，
下载block失败的情况添加进去了
1)用Process之间的Condition替换Semaphore，用SystemRandom替换全局的random
2)替换旧的broadcast，删掉所有旧的broadcast
3)显示运行过程中的failed log

<5>fix: fix the bug of report bad
将download失败的相关log设为warning

<6>fix: impl the store blocks function by mmap

1)用mmap实现blocks的管理与存储
2)将原先的bitmap弄的更详细一点，可以显示broadcast的每个块的offset, size
3)原来的mmap出现了一些问题，原因是因为同一个节点上可能同时驻留两个downloadManager
需要区分存储到本地的文件
4)之前在测试运行程序的时候，出现了corrupt input或者chk不对的情况，经过确认
主要原因是读写的时候，没有加锁，导致出现错误
5)将lock去掉，对mmap，分别使用不同类型的mmap，区分读与写，实现无锁状态

<7>fix: Adjust the multiprocess scheduler for new broadcast

在新broadcast实现的时候，需要调整原来的multiprocess的实现方式
1)将pool的初始化放在submit的时候进行，
如果在一开始就执行初始化，会导致后来的broadcast
环境变量的配置在进程中不可见，multiprocessing的manager被hung住
2)pool的初始化放在之后的话，会继承在context中配置的信号处理函数
这样在终止broadcast时，会出现信号处理的日志，现在的逻辑是将信号处理的
函数进行还原。

+++ dpark/schedule.py
@@ -6,6 +6,7 @@ import multiprocessing
 import os
 import random
 import socket
+import signal
 import sys
 import time
 from six.moves import urllib
@@ -33,7 +34,7 @@ from six.moves import range
 logger = get_logger(__name__)
 
 EXECUTOR_CPUS = 0.01
-EXECUTOR_MEMORY = 64  # cache
+EXECUTOR_MEMORY = 128 # cache
 POLL_TIMEOUT = 0.1
 RESUBMIT_TIMEOUT = 60
 MAX_IDLE_TIME = 60 * 30
@@ -481,12 +482,22 @@ class LocalScheduler(DAGScheduler):
             self.taskEnded(task, reason, result, update)
 
 
-def run_task_in_process(task, tid, environ):
+def run_task_in_process(task, tid, environ, sig_dict=None):
     from dpark.env import env
     workdir = environ.get('WORKDIR')
     environ['SERVER_URI'] = 'file://%s' % workdir[0]
+    environ['broadcast_task'] = 1
     env.start(False, environ)
-
+    import signal
+    if sig_dict:
+        if 'SIGTERM' in sig_dict:
+            signal.signal(signal.SIGTERM, sig_dict['SIGTERM'])
+        if 'SIGHUP' in sig_dict:
+            signal.signal(signal.SIGHUP, sig_dict['SIGHUP'])
+        if 'SIGABRT' in sig_dict:
+            signal.signal(signal.SIGABRT, sig_dict['SIGABRT'])
+        if 'SIGQUIT' in sig_dict:
+            signal.signal(signal.SIGQUIT, sig_dict['SIGQUIT'])
     logger.debug('run task in process %s %s', task, tid)
     try:
         return run_task(task, tid)
@@ -500,7 +511,11 @@ class MultiProcessScheduler(LocalScheduler):
         LocalScheduler.__init__(self)
         self.threads = threads
         self.tasks = {}
-        self.pool = multiprocessing.Pool(self.threads or 2)
+        self.pool = None
+        self.sig_dict = {'SIGTERM': signal.getsignal(signal.SIGTERM),
+                         'SIGHUP': signal.getsignal(signal.SIGHUP),
+                         'SIGABRT': signal.getsignal(signal.SIGABRT),
+                         'SIGQUIT': signal.getsignal(signal.SIGQUIT)}
 
     def submitTasks(self, tasks):
         if not tasks:
@@ -526,13 +541,16 @@ class MultiProcessScheduler(LocalScheduler):
         for task in tasks:
             logger.debug('put task async: %s', task)
             self.tasks[task.id] = task
+            if not self.pool:
+                self.pool = multiprocessing.Pool(self.threads or 2)
             self.pool.apply_async(run_task_in_process,
-                                  [task, self.nextAttempId(), env.environ],
+                                  [task, self.nextAttempId(), env.environ, self.sig_dict],
                                   callback=callback)
 
     def stop(self):
-        self.pool.terminate()
-        self.pool.join()
+        if self.pool:
+            self.pool.terminate()
+            self.pool.join()
         logger.debug('process pool stopped')
